Add rudimentary auto-config option
Add a UI option to automtaically configure the reference length and penis
base offset based on a duration-configureable history of penis utilization
stats.
Eg if 50% of the penis is being used over the course of 3 seconds, adjust
the length to 50% and adjust the penis base offset to almost intersect with
the lowest thrust depth.